requirements.txt
flask~=3.0.0
audiocraft~=1.0.0
openai~=1.3.0
requests~=2.31.0
python-dotenv~=1.0.0
Pillow~=9.3.0
torch~=2.0.1+cu117
torchaudio~=2.0.2+cu117
line-bot-sdk~=3.5.1
pyngrok~=7.0.1